Enable caching of input and output of the gradle task
Summary: Outputs of the task (there are none) are now explicitly considered always up-to-date and inputs have been made serializable. This way we can use the caching of the task result and avoid running the task if inputs are unchanged.

Test Plan: Manually tested the upload to AS

Reviewers: samuel.hopstock

Reviewed By: samuel.hopstock

Maniphest Tasks: T15679

Differential Revision: https://phabricator.guardsquare.com/D15451