Use GitHub SHA value if provided.
Summary:
Based on T20679, when a `GITHUB_SHA` value is provided this gradle-plugin should not run a `git` command to get the hash value of the latest commit. Instead, it should use provided `GITHUB_SHA` value.

To achieve this, the gradle-plugin code changed to check for the environment variable `GITHUB_SHA`. In case it is provided, it will be used as a commit hash, and it will not run any git command to obtain the hash value of the latest commit.

Closes T20679.

Test Plan: Test locally to check if a dummy commit hash value is appeared on https://appsweep.guardsquare.com

Reviewers: brent.chesny, dennis.titze

Reviewed By: brent.chesny, dennis.titze

Subscribers: matthijs.kaminski, ryan.lloyd, dennis.titze

Maniphest Tasks: T20679

Differential Revision: https://phabricator.guardsquare.com/D19743