Change README to avoid confusion of creating API Key.
Summary:
In the configuration section, we are talking about configuring the project, and then suddenly a small point mentions creating an API key for the project. It brought this idea that we need to do this part in Android Studio, because, everything till this point was done in Android Studio, but we need to create an API key in AppSweep project setting.

In this diff, two options were suggested to change the description.

PS Please don't land this diff. One option needs to be removed.

Test Plan: NA

Reviewers: dennis.titze

Reviewed By: dennis.titze

Differential Revision: https://phabricator.guardsquare.com/D19818